[CHANGE] Cleanup contacts list, add online indicator
