+++ pandora-client-web/src/components/releationships/relationships.tsx
@@ -1,7 +1,6 @@
 import React, { useMemo } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { AccountId, AsyncSynchronized, IAccountFriendStatus, IAccountRelationship, IClientDirectory, IConnectionBase, IDirectoryClientArgument, TypedEventEmitter } from 'pandora-common';
-import { Observable, useObservable } from '../../observable';
+import { AccountId, IAccountFriendStatus, IAccountRelationship } from 'pandora-common';
 import { Tab, TabContainer } from '../common/tabs/tabs';
 import { DirectMessages } from '../directMessages/directMessages';
 import './relationships.scss';
@@ -9,80 +8,9 @@ import { Button } from '../common/button/button';
 import { useDirectoryConnector } from '../gameContext/directoryConnectorContextProvider';
 import { NotificationSource, useNotificationSuppressed } from '../gameContext/notificationContextProvider';
 import { useAsyncEvent } from '../../common/useEvent';
-import { toast } from 'react-toastify';
-import { TOAST_OPTIONS_ERROR } from '../../persistentToast';
 import _ from 'lodash';
-
-const RELATIONSHIPS = new Observable<readonly IAccountRelationship[]>([]);
-const FRIEND_STATUS = new Observable<readonly IAccountFriendStatus[]>([]);
-
-export const RelationshipContext = new class RelationshipContext extends TypedEventEmitter<{
-	incoming: IAccountRelationship & { type: 'incoming'; };
-}> {
-	private _queue: (() => void)[] = [];
-	private _useQueue = true;
-
-	@AsyncSynchronized()
-	public async initStatus(connection: IConnectionBase<IClientDirectory>): Promise<void> {
-		if (!this._useQueue) {
-			return;
-		}
-		const { friends, relationships } = await connection.awaitResponse('getRelationships', {});
-		RELATIONSHIPS.value = relationships;
-		FRIEND_STATUS.value = friends;
-		this._dequeue();
-	}
-
-	public handleFriendStatus(data: IDirectoryClientArgument['friendStatus']) {
-		if (this._useQueue) {
-			this._queue.push(() => this.handleFriendStatus(data));
-			return;
-		}
-		const filtered = FRIEND_STATUS.value.filter((status) => status.id !== data.id);
-		if (data.online !== 'delete') {
-			filtered.push(data);
-		}
-		FRIEND_STATUS.value = filtered;
-	}
-
-	public handleRelationshipsUpdate({ relationship, friendStatus }: IDirectoryClientArgument['relationshipsUpdate']) {
-		if (this._useQueue) {
-			this._queue.push(() => this.handleRelationshipsUpdate({ relationship, friendStatus }));
-			return;
-		}
-		// Update relationship side
-		{
-			const filtered = RELATIONSHIPS.value.filter((currentRelationship) => currentRelationship.id !== relationship.id);
-			if (relationship.type !== 'none') {
-				filtered.push(relationship);
-				if (filtered.length > RELATIONSHIPS.value.length && relationship.type === 'incoming') {
-					this.emit('incoming', { ...relationship, type: 'incoming' });
-				}
-			}
-			RELATIONSHIPS.value = filtered;
-		}
-		// Update friend side
-		{
-			const filtered = FRIEND_STATUS.value.filter((status) => status.id !== friendStatus.id);
-			if (friendStatus.online !== 'delete') {
-				filtered.push(friendStatus);
-			}
-			FRIEND_STATUS.value = filtered;
-		}
-	}
-
-	public handleLogout() {
-		this._useQueue = true;
-		FRIEND_STATUS.value = [];
-		RELATIONSHIPS.value = [];
-	}
-
-	private _dequeue() {
-		this._useQueue = false;
-		this._queue.forEach((fn) => fn());
-		this._queue = [];
-	}
-};
+import { Row } from '../common/container/container';
+import { RelationshipChangeHandleResult, useFriendStatus, useRelationships } from './relationshipsContext';
 
 export function Relationships() {
 	const navigate = useNavigate();
@@ -110,16 +38,6 @@ export function Relationships() {
 			</TabContainer>
 		</div>
 	);
-}
-
-export function useRelationships(type: IAccountRelationship['type']) {
-	const rel = useObservable(RELATIONSHIPS);
-	return useMemo(() => rel.filter((r) => r.type === type), [rel, type]);
-}
-
-export function useRelationship(id: AccountId): IAccountRelationship | undefined {
-	const rel = useObservable(RELATIONSHIPS);
-	return useMemo(() => rel.find((r) => r.id === id), [rel, id]);
 }
 
 function RelationshipHeader({ type }: { type: IAccountRelationship['type']; }) {
@@ -202,7 +120,7 @@ function PendingRequestActions({ id }: { id: AccountId; }) {
 	const directory = useDirectoryConnector();
 	const [cancel, cancelInProgress] = useAsyncEvent(async () => {
 		return await directory.awaitResponse('friendRequest', { id, action: 'cancel' });
-	}, (r) => HandleResult(r?.result));
+	}, (r) => RelationshipChangeHandleResult(r?.result));
 	return (
 		<Button className='slim' onClick={ cancel } disabled={ cancelInProgress }>Cancel</Button>
 	);
@@ -215,13 +133,13 @@ function IncomingRequestActions({ id }: { id: AccountId; }) {
 			return await directory.awaitResponse('friendRequest', { id, action: 'accept' });
 		}
 		return undefined;
-	}, (r) => HandleResult(r?.result));
+	}, (r) => RelationshipChangeHandleResult(r?.result));
 	const [decline, declineInProgress] = useAsyncEvent(async () => {
 		if (confirm(`Decline the request to add ${id} to your contacts?`)) {
 			return await directory.awaitResponse('friendRequest', { id, action: 'decline' });
 		}
 		return undefined;
-	}, (r) => HandleResult(r?.result));
+	}, (r) => RelationshipChangeHandleResult(r?.result));
 	return (
 		<>
 			<Button className='slim' onClick={ accept } disabled={ acceptInProgress }>Accept</Button>
@@ -232,7 +150,7 @@ function IncomingRequestActions({ id }: { id: AccountId; }) {
 
 function ShowFriends() {
 	const friends = useRelationships('friend');
-	const status = useObservable(FRIEND_STATUS);
+	const status = useFriendStatus();
 	const friendsWithStatus = useMemo(() => {
 		return friends.map((friend) => {
 			const stat = status.find((s) => s.id === friend.id);
@@ -240,7 +158,7 @@ function ShowFriends() {
 				id: friend.id,
 				name: friend.name,
 				time: friend.time,
-				status: stat?.online ? 'online' : 'offline',
+				online: stat?.online === true,
 				characters: stat?.characters,
 			};
 		});
@@ -248,13 +166,21 @@ function ShowFriends() {
 
 	return (
 		<table>
+			<colgroup>
+				<col style={ { width: '1%' } } />
+				<col />
+				<col />
+				<col />
+				<col style={ { width: '1%' } } />
+				<col style={ { width: '1%' } } />
+			</colgroup>
 			<thead>
 				<tr>
 					<th>ID</th>
 					<th>Name</th>
-					<th>Created</th>
 					<th>Status</th>
 					<th>Online Characters</th>
+					<th>Since</th>
 					<th>Actions</th>
 				</tr>
 			</thead>
@@ -271,13 +197,13 @@ function FriendRow({
 	id,
 	name,
 	time,
-	status,
+	online,
 	characters,
 }: {
 	id: AccountId;
 	name: string;
 	time: number;
-	status: string;
+	online: boolean;
 	characters?: IAccountFriendStatus['characters'];
 }) {
 	const directory = useDirectoryConnector();
@@ -287,15 +213,22 @@ function FriendRow({
 			return await directory.awaitResponse('unfriend', { id });
 		}
 		return undefined;
-	}, (r) => HandleResult(r?.result));
+	}, (r) => RelationshipChangeHandleResult(r?.result));
 
 	return (
-		<tr>
-			<td>{ id }</td>
-			<td>{ name }</td>
-			<td>{ new Date(time).toLocaleString() }</td>
-			<td>{ status }</td>
+		<tr className={ online ? 'friend online' : 'friend offline' }>
+			<td className='selectable'>{ id }</td>
+			<td className='selectable'>{ name }</td>
+			<td className='status'>
+				<Row className='fill' alignX='center' alignY='center'>
+					<span className='indicator'>
+						{ online ? '\u25CF' : '\u25CB' }
+					</span>
+					{ online ? 'Online' : 'Offline' }
+				</Row>
+			</td>
 			<td>{ characters?.map((c) => c.name).join(', ') }</td>
+			<td>{ new Date(time).toLocaleDateString() }</td>
 			<td>
 				<Button className='slim' onClick={ unfriend } disabled={ processing }>
 					Remove
@@ -303,27 +236,4 @@ function FriendRow({
 			</td>
 		</tr>
 	);
-}
-
-export function HandleResult(result: 'ok' | 'accountNotFound' | 'requestNotFound' | 'blocked' | 'requestAlreadyExists' | undefined) {
-	switch (result) {
-		case undefined:
-		case 'ok':
-			return;
-		case 'accountNotFound':
-			toast('Account not found', TOAST_OPTIONS_ERROR);
-			return;
-		case 'requestNotFound':
-			toast('Request not found', TOAST_OPTIONS_ERROR);
-			return;
-		case 'blocked':
-			toast('Account is blocked', TOAST_OPTIONS_ERROR);
-			return;
-		case 'requestAlreadyExists':
-			toast('Request already exists', TOAST_OPTIONS_ERROR);
-			return;
-		default:
-			toast('Unknown error', TOAST_OPTIONS_ERROR);
-			return;
-	}
 }
